+++ packages/freezed/lib/src/freezed_generator.dart
@@ -61,6 +61,8 @@ class ConstructorDetails {
     @required this.impliedProperties,
     @required this.fullName,
     @required this.decorators,
+    @required this.withDecorators,
+    @required this.implementsDecorators,
     @required this.hasJsonSerializable,
     @required this.cloneableProperties,
     @required this.canOverrideToString,
@@ -74,6 +76,8 @@ class ConstructorDetails {
   final bool isDefault;
   final bool hasJsonSerializable;
   final String fullName;
+  final List<String> withDecorators;
+  final List<String> implementsDecorators;
   final List<String> decorators;
   final List<CloneableProperty> cloneableProperties;
   final bool canOverrideToString;
@@ -92,6 +96,8 @@ $runtimeType(
   impliedProperties: $impliedProperties,
   fullName: $fullName,
   decorators: $decorators,
+  withDecorators: $withDecorators,
+  implementsDecorators: $implementsDecorators,
   hasJsonSerializable: $hasJsonSerializable,
   cloneableProperties: $cloneableProperties,
   canOverrideToString: $canOverrideToString,
@@ -393,6 +399,34 @@ class FreezedGenerator extends ParserGenerator<_GlobalData, Data, Freezed> {
         }
       }
 
+      Iterable<String> withDecorationTypes() sync* {
+        for (final metadata in constructor.metadata) {
+          if (!metadata.isWith) continue;
+          final object = metadata.computeConstantValue();
+          var type = object.getField('type');
+          if (type.isNull) {
+            type = object.getField('stringType');
+            yield type.toStringValue();
+          } else {
+            yield type.toTypeValue().getDisplayString();
+          }
+        }
+      }
+
+      Iterable<String> implementsDecorationTypes() sync* {
+        for (final metadata in constructor.metadata) {
+          if (!metadata.isImplements) continue;
+          final object = metadata.computeConstantValue();
+          var type = object.getField('type');
+          if (type.isNull) {
+            type = object.getField('stringType');
+            yield type.toStringValue();
+          } else {
+            yield type.toTypeValue().getDisplayString();
+          }
+        }
+      }
+
       result.add(
         ConstructorDetails(
           name: constructor.name,
@@ -404,6 +438,8 @@ class FreezedGenerator extends ParserGenerator<_GlobalData, Data, Freezed> {
               Property.fromParameter(parameter),
           ],
           decorators: constructor.metadata.map((e) => e.toSource()).toList(),
+          withDecorators: withDecorationTypes().toSet().toList(),
+          implementsDecorators: implementsDecorationTypes().toSet().toList(),
           isDefault: isDefaultConstructor(constructor),
           hasJsonSerializable: constructor.hasJsonSerializable,
           cloneableProperties: cloneableProperties().toList(),